feat: use GL 3.3

+++ include/Netycat/include/Netycat/Core/Socket.h
@@ -43,7 +43,7 @@ namespace Netycat {
             
             private:
             
-            int sock;
+            long long sock;
             InetAddress* address;
             
             Buffer recvBuffer;
